Added two links to explain reproducibility and to the Git repo
Until now there wasn't any explanation of what we mean by "exact
reproducibility", so I added it and also put a click-able link to the
Git repository for easy navigation.

This is indeed just a crude place-holder, but since its live now, and
we have mentioned it in the draft paper, its good to have some
links. It also helps web search tools to already connect this page
with other related pages so by the time the final page is ready, it
can already have a presence in the search results.