Tidy up the CSS file, esp. the colors section
Created a seperate color section at the beginning of the CSS file, so
one can change colors without moving around the CSS file so much!

Created a `FONTS` section at the top of the CSS file.

Also, moved the `decoration classes` section next to the `colors`
section in the CSS file.

I'll try to tidy up the CSS file further.