New project-flow SVG image, new .center CSS class
Until now, The project-flow SVG image was saved as optimized SVG
with comments saved in it by Inkscape. This made its size about
than 230 KB. The icon of paper in the image was also a simple paper.

With this commit, I modified the icon of the paper and added some
details: a small shape made of a square and circle inside the
text of the paper to look like a figure. Then I saved it as an
optimized SVG file without any comments. This allows us to have
a file with 208KB size.

For putting the image in the website I also decreased its size, so
it was necessary to center it. To do this, I added a general class
'.center' for this job and added it in 'base.css' at the end of
General style section to allow us have everything with this class
in the center.