Added an image to link to RDA talk on Youtube
Until now, the raw link as text wasn't too obvious for a new viewer of
the webpage. So a snapshot was taken from the start of the video
created by RDA of my talk and added in the page to be more prominent
and help new viewers notice it more clearly.