Figure added for showing file architecture
In the initial version of the paper, we had this nice figure to show
the file structure of Maneage, so I thought its good to put it in the
page describing Maneage's file architecture.

Also, the long copyright at the end of the page was summarized into
one line and a link to the GPL has been added.