Update refs; add codeberg to repository list
This commit updates the list of references for
Maneage publications; and modifies the comment
in about-tips.html that listed the unethical servers
without listing ethical servers. The current best
list of ethical git (and similar) code hosts seems to
be at  https://developers.reverseeagle.org/replace/github .

(The equivalent for non-developers is

    https://switching.software

to avoid the closed world of GAFAM/Zoom/Slack/Doodle/Twitter
and join the free world. :))