Minor edits in citation page
Following the previous commit, I noticed that in the first line of
that page (where we ask people to cite the Maneage paper), don't
actually mention the journal or DOI. So a link with the journal name,
that links to the DOI was added.

I also noticed that the Peper & Roukema paper's new information
doesn't follow the same convention as the previous entries: instead of
the year being set as a link, the journal name was a link. So the link
is now on the year.

I also noticed that there was a typo in the link of the Maneage
paper's arXiv link that has been corrected.