Unified structure of papers using Maneage
The top few entries in the list were different from the older ones. So
for now (since I am busy!), they have been written in a similar format
(in particular, removing the arXiv link).

I personally prefer the new method, but we also need consistency! We
can't just update the most recent entries and ignore the previous
ones: this doesn't have a good view for a random viewer.

So when we have time, we should define some good way to add those
extra links in all the papers ;-).