Added CSS, allowing iframe aspect ratio to be preserved
Until now, a fixed height of 315 was assumed for the 'iframe'. So on a
full screen monitor the video would be too thin and on a smarphone it
would be too thick.

With this commit, I found a good CSS style that conforms to CSS2 for
older browsers also and used it to have varying aspect ratios based on
the size of the browser.

Also, now that people can directly watch the video on Maneage's own
webpage, it was necessary to edit the text above the video.