Added notice for PeerTube video not loading
Until now, when http://maneage.org was used, the video would not
load. However, when https://maneage.org was used (with a HTTPS), the
video would load. This is because PeerTube uses HTTPS. Boud has
provided a very complete description of the problem in [1].

With this commit, until we find a good/robust solution, a notice has
been added just under the video (in a slightly smaller font), hinting
at the solution for people who can't play the video.

[1] https://savannah.nongnu.org/task/?15654#comment12