Customized project branch suggestion is now main (following Git)
Until now in the main webpage, we were suggesting to call the new
branch (to start customizing) as 'master' (following the old Git
convention). However, Git has recently changed its default branch name
to 'main'.

With this commit, to be consistent with other newly created Git repos,
we are suggesting to call the main customized branch as 'main'.