Forum: add the support and forum page, wip
Until now, Maneage users could ask for help via the savannah page and
direct email.

With this commit, a new forum page is added to the website that encourages
user/developers to join the Maneage room on open astronomy.

The actual room is not created yet, but this seems a starting point.